Add hostname for correct Sauce Connect functioning.
Sauce Connect is the traffic proxy tool for Sauce Connect tests.

On some test platforms, the OS ignored configured proxies for localhost.
This prevents the browser from connecting to the karma server.

By using a hostname instead of localhost, host resolution is done
on the client side (eg Travis) and the problem is resolved.

This change sets the hostname as 'karmalocal' and instructs Travis to use the same.